requirements.txt
openai
python-dotenv
requests


dependency-injector==4.48.3
